[AN] refactor: 홈 화면 어댑터 패키지 구조 변경 #711
*   [AN] refactor: 홈 화면 관련 어댑터들을 `adapter` 하위 패키지로 이동
