+++ android/app/src/main/java/com/onair/hearit/presentation/home/adapter/PlayingHistoryHearitViewHolder.kt
@@ -1,15 +1,15 @@
-package com.onair.hearit.presentation.home
+package com.onair.hearit.presentation.home.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.onair.hearit.analytics.HearitSource
-import com.onair.hearit.databinding.ItemRecentUploadHearitBinding
-import com.onair.hearit.domain.model.RecentUploadHearit
+import com.onair.hearit.databinding.ItemPlayingHistoryHearitBinding
+import com.onair.hearit.domain.model.PlayingHistoryHearit
 import com.onair.hearit.presentation.HearitClickListener
 
-class RecentUploadHearitViewHolder private constructor(
-    private val binding: ItemRecentUploadHearitBinding,
+class PlayingHistoryHearitViewHolder private constructor(
+    private val binding: ItemPlayingHistoryHearitBinding,
     private val source: HearitSource,
     hearitClickListener: HearitClickListener,
 ) : RecyclerView.ViewHolder(binding.root) {
@@ -18,8 +18,8 @@ class RecentUploadHearitViewHolder private constructor(
         binding.source = source
     }
 
-    fun bind(item: RecentUploadHearit) {
-        binding.recentUploadHearit = item
+    fun bind(item: PlayingHistoryHearit) {
+        binding.playingHistoryHearit = item
     }
 
     companion object {
@@ -27,10 +27,10 @@ class RecentUploadHearitViewHolder private constructor(
             parent: ViewGroup,
             source: HearitSource,
             hearitClickListener: HearitClickListener,
-        ): RecentUploadHearitViewHolder {
+        ): PlayingHistoryHearitViewHolder {
             val inflater = LayoutInflater.from(parent.context)
-            val binding = ItemRecentUploadHearitBinding.inflate(inflater, parent, false)
-            return RecentUploadHearitViewHolder(binding, source, hearitClickListener)
+            val binding = ItemPlayingHistoryHearitBinding.inflate(inflater, parent, false)
+            return PlayingHistoryHearitViewHolder(binding, source, hearitClickListener)
         }
     }
 }
