+++ android/app/src/main/java/com/onair/hearit/presentation/home/adapter/RecentUploadHearitViewHolder.kt
@@ -1,25 +1,25 @@
-package com.onair.hearit.presentation.home
+package com.onair.hearit.presentation.home.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.onair.hearit.analytics.HearitSource
-import com.onair.hearit.databinding.ItemRecommendHearitBinding
-import com.onair.hearit.domain.model.RecommendHearit
+import com.onair.hearit.databinding.ItemRecentUploadHearitBinding
+import com.onair.hearit.domain.model.RecentUploadHearit
 import com.onair.hearit.presentation.HearitClickListener
 
-class RecommendViewHolder(
-    private val binding: ItemRecommendHearitBinding,
+class RecentUploadHearitViewHolder private constructor(
+    private val binding: ItemRecentUploadHearitBinding,
     private val source: HearitSource,
     hearitClickListener: HearitClickListener,
 ) : RecyclerView.ViewHolder(binding.root) {
     init {
-        binding.recommendClickListener = hearitClickListener
+        binding.clickListener = hearitClickListener
         binding.source = source
     }
 
-    fun bind(item: RecommendHearit) {
-        binding.item = item
+    fun bind(item: RecentUploadHearit) {
+        binding.recentUploadHearit = item
     }
 
     companion object {
@@ -27,10 +27,10 @@ class RecommendViewHolder(
             parent: ViewGroup,
             source: HearitSource,
             hearitClickListener: HearitClickListener,
-        ): RecommendViewHolder {
+        ): RecentUploadHearitViewHolder {
             val inflater = LayoutInflater.from(parent.context)
-            val binding = ItemRecommendHearitBinding.inflate(inflater, parent, false)
-            return RecommendViewHolder(binding, source, hearitClickListener)
+            val binding = ItemRecentUploadHearitBinding.inflate(inflater, parent, false)
+            return RecentUploadHearitViewHolder(binding, source, hearitClickListener)
         }
     }
 }
