+++ android/app/src/main/java/com/onair/hearit/presentation/home/adapter/RecommendHearitAdapter.kt
@@ -1,36 +1,36 @@
-package com.onair.hearit.presentation.home
+package com.onair.hearit.presentation.home.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.onair.hearit.analytics.HearitSource
-import com.onair.hearit.domain.model.RecentUploadHearit
+import com.onair.hearit.domain.model.RecommendHearit
 import com.onair.hearit.presentation.HearitClickListener
 
-class RecentUploadHearitAdapter(
+class RecommendHearitAdapter(
     private val hearitClickListener: HearitClickListener,
-) : ListAdapter<RecentUploadHearit, RecentUploadHearitViewHolder>(DiffCallback) {
+) : ListAdapter<RecommendHearit, RecommendViewHolder>(DiffCallback) {
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int,
-    ): RecentUploadHearitViewHolder = RecentUploadHearitViewHolder.create(parent, HearitSource.RECENT_UPLOAD, hearitClickListener)
+    ): RecommendViewHolder = RecommendViewHolder.create(parent, HearitSource.RECOMMEND, hearitClickListener)
 
     override fun onBindViewHolder(
-        holder: RecentUploadHearitViewHolder,
+        holder: RecommendViewHolder,
         position: Int,
     ) = holder.bind(getItem(position))
 
     companion object {
-        private val DiffCallback =
-            object : DiffUtil.ItemCallback<RecentUploadHearit>() {
+        val DiffCallback =
+            object : DiffUtil.ItemCallback<RecommendHearit>() {
                 override fun areItemsTheSame(
-                    oldItem: RecentUploadHearit,
-                    newItem: RecentUploadHearit,
+                    oldItem: RecommendHearit,
+                    newItem: RecommendHearit,
                 ): Boolean = oldItem.id == newItem.id
 
                 override fun areContentsTheSame(
-                    oldItem: RecentUploadHearit,
-                    newItem: RecentUploadHearit,
+                    oldItem: RecommendHearit,
+                    newItem: RecommendHearit,
                 ): Boolean = oldItem == newItem
             }
     }
