+++ android/app/src/main/java/com/onair/hearit/presentation/home/adapter/RecommendViewHolder.kt
@@ -1,25 +1,25 @@
-package com.onair.hearit.presentation.home
+package com.onair.hearit.presentation.home.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.onair.hearit.analytics.HearitSource
-import com.onair.hearit.databinding.ItemPlayingHistoryHearitBinding
-import com.onair.hearit.domain.model.PlayingHistoryHearit
+import com.onair.hearit.databinding.ItemRecommendHearitBinding
+import com.onair.hearit.domain.model.RecommendHearit
 import com.onair.hearit.presentation.HearitClickListener
 
-class PlayingHistoryHearitViewHolder private constructor(
-    private val binding: ItemPlayingHistoryHearitBinding,
+class RecommendViewHolder(
+    private val binding: ItemRecommendHearitBinding,
     private val source: HearitSource,
     hearitClickListener: HearitClickListener,
 ) : RecyclerView.ViewHolder(binding.root) {
     init {
-        binding.clickListener = hearitClickListener
+        binding.recommendClickListener = hearitClickListener
         binding.source = source
     }
 
-    fun bind(item: PlayingHistoryHearit) {
-        binding.playingHistoryHearit = item
+    fun bind(item: RecommendHearit) {
+        binding.item = item
     }
 
     companion object {
@@ -27,10 +27,10 @@ class PlayingHistoryHearitViewHolder private constructor(
             parent: ViewGroup,
             source: HearitSource,
             hearitClickListener: HearitClickListener,
-        ): PlayingHistoryHearitViewHolder {
+        ): RecommendViewHolder {
             val inflater = LayoutInflater.from(parent.context)
-            val binding = ItemPlayingHistoryHearitBinding.inflate(inflater, parent, false)
-            return PlayingHistoryHearitViewHolder(binding, source, hearitClickListener)
+            val binding = ItemRecommendHearitBinding.inflate(inflater, parent, false)
+            return RecommendViewHolder(binding, source, hearitClickListener)
         }
     }
 }
